digital-twin-chat/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.0
boto3==1.35.0
mangum==0.18.0
openai==1.54.0
hypothesis==6.115.0
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
pypdf2==3.0.1
